+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 ### 0.9.0
 * [EventStore][Breaking] Move StreamVersion to the Events namespace.
 * [EventStore][Breaking] Rename EventStreamPosition to EventStreamHeader.
+* [EventStore][Improvment] Partitioned notificaion channel.
 
 ### 0.8.5
 * [EventStore][Improvment] Caching data in EventJournalReaders and EventStreamConsumers.